+++ rust-src/wallet_library/CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Added
 
 - Added method `serialize_credential_deployment_details` for serializing the credential deployment details that must be signed to deploy a new credential.
+- Added method `serialize_credential_deployment_payload` for serializing the credential deployment payload. 
+This payload is the one sent as a raw payload to the node when deploying a new credential.
 
 ## [0.1.0]